AsyncKeyedLock: Rename folder and add link to the original source code.
