+++ src/MudBlazor/Utilities/AsyncKeyedLock/AsyncKeyedLockTimeoutReleaser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2023 Mark Cilia Vincenti
+﻿// Copyright (c) 2023 Mark Cilia Vincenti (original source code: https://github.com/MarkCiliaVincenti/AsyncKeyedLock)
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -6,7 +6,7 @@ using System;
 using System.Runtime.CompilerServices;
 using System.Threading;
 
-namespace MudBlazor.Utilities.AsyncKeyedLocker;
+namespace MudBlazor.Utilities.AsyncKeyedLock;
 
 #nullable enable
 /// <summary>
